requirements.txt
# CoderCombat Tournament Management System Dependencies

# Database (choose one; remove the other if unused)
pymysql>=1.0.2
# mysql-connector-python>=8.0.33

# HTTP requests for DOMjudge API
requests>=2.28.0

# Load environment variables from .env file
python-dotenv>=1.0.0

# Development dependencies (optional)
# pytest>=7.0.0   # For testing
# black>=22.0.0   # Code formatting
# flake8>=4.0.0   # Linting
